# product-service/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:9002
  description: Generated server url
paths:
  /saveProduct:
    post:
      tags:
      - product-controller
      operationId: saveProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductDTO'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductDTO'
  /products:
    get:
      tags:
      - product-controller
      operationId: getAllProductsWithParameter
      parameters:
      - name: flag
        in: query
        required: true
        schema:
          type: boolean
      - name: appCallerName
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product:
    get:
      tags:
      - product-controller
      operationId: getAllProducts
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/{flag}:
    get:
      tags:
      - product-controller
      operationId: getAllProductsWithFlag
      parameters:
      - name: flag
        in: path
        required: true
        schema:
          type: boolean
      - name: appCallerName
        in: header
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
  /product/parameter:
    get:
      tags:
      - product-controller
      operationId: getAllProductsWithParam
      parameters:
      - name: tokens
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ProductDTO'
components:
  schemas:
    ProductDTO:
      type: object
      properties:
        productId:
          type: string
        productName:
          type: string
        price:
          type: integer
          format: int64
        stock:
          type: integer
          format: int32
        productType:
          type: string
        port:
          type: integer
          format: int32