# product-service/src/main/java/com/mitocode/microservices/productservice/model/dto/ProductDTO.java
package com.mitocode.microservices.productservice.model.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.*;

//@Data
@Getter
@Setter
@Builder
@AllArgsConstructor
@NoArgsConstructor
public class ProductDTO {

    private String productId;
    @NotBlank
    private String productName;
    private Long price;
    private Integer stock;
    private String productType;
    private Integer port;


}